+++ packages/zosfiles/src/constants/ZosFiles.messages.ts
@@ -208,15 +208,6 @@ export const ZosFilesMessages: { [key: string]: IMessageDefinition } = {
     },
 
     /**
-     * Message indicating that the data sets matching pattern was downloaded successfully
-     * @type {IMessageDefinition}
-     * @deprecated
-     */
-    datasetsDownloadedSuccessfully: {
-        message: "Data sets matching pattern downloaded successfully.\nDestination: %s"
-    },
-
-    /**
      * Message indicating that the data sets matching pattern were listed successfully
      * @type {IMessageDefinition}
      */
@@ -409,24 +400,6 @@ export const ZosFilesMessages: { [key: string]: IMessageDefinition } = {
     },
 
     /**
-     * Message indicating that all data sets matching the provided patterns are archived.
-     * @type {IMessageDefinition}
-     * @deprecated
-     */
-    allDataSetsArchived: {
-        message: "All data sets matching the selected pattern(s) were archived."
-    },
-
-    /**
-     * Message indicating that no data sets remain to be downloaded after the excluded ones were filtered out.
-     * @type {IMessageDefinition}
-     * @deprecated
-     */
-    noDataSetsInList: {
-        message: "No data sets left after excluded pattern(s) were filtered out."
-    },
-
-    /**
      * Message indicating that no data sets remain to be downloaded after the excluded ones were filtered out.
      * @type {IMessageDefinition}
      */
@@ -437,20 +410,9 @@ export const ZosFilesMessages: { [key: string]: IMessageDefinition } = {
     /**
      * Message indicating that no data sets remain to be downloaded after the excluded ones were filtered out.
      * @type {IMessageDefinition}
-     * @deprecated
      */
-    noDataSetsMatchingPatternRemain: {
-        message: "After filtering out the archived files and files that match the exclusion-parameters, no data sets matching" +
-        " the supported organization type remain."
-    },
-
-    /**
-     * Message indicating that only empty partitioned data sets match the provided patterns
-     * @type {IMessageDefinition}
-     * @deprecated
-     */
-    onlyEmptyPartitionedDataSets: {
-        message: "Only empty partitioned data sets match the provided patterns."
+    noDataSetsInList: {
+        message: "No data sets left after excluded pattern(s) were filtered out."
     },
 
     /**